requirements.txt
ccxt
begins
